Moved UnsupportedPyArrowTypeException to pyarrow.py
